move Numeric tests into own test class
also: implicit casting INT and LONG constants to long256
